pyproject.toml
[tool.poetry]
name = "document-processing"
version = "0.1.0"
description = ""
authors = ["Steve Ahlswede <steve.ahlswede@gsk.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pymupdf = "^1.24.14"
black = "^24.10.0"
mypy = "^1.13.0"
llama-index = "^0.12.1"
llama-index-embeddings-azure-openai = "^0.3.0"
python-docx = "^1.1.2"
isort = "^5.13.2"
pymupdf4llm = "^0.0.17"
langdetect = "^1.0.9"
pytest = "^8.3.3"
docx2python = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"